refactor(CarView): CarView를 CarInputView와 CarOutputView로 분리
